more separation of AvitoReports
