+++ subprojects/common/report-viewer/src/main/kotlin/com/avito/report/ReportLinkGenerator.kt
@@ -9,10 +9,13 @@ interface ReportLinkGenerator {
 
     fun generateTestLink(testName: TestName): String
 
-    object Stub : ReportLinkGenerator {
+    class Stub(
+        private val reportLink: String = "",
+        private val testLink: String = ""
+    ) : ReportLinkGenerator {
 
-        override fun generateReportLink(filterOnlyFailtures: Boolean, team: Team): String = ""
+        override fun generateReportLink(filterOnlyFailtures: Boolean, team: Team): String = reportLink
 
-        override fun generateTestLink(testName: TestName): String = ""
+        override fun generateTestLink(testName: TestName): String = testLink
     }
 }
