+++ subprojects/gradle/instrumentation-tests/src/main/kotlin/com/avito/instrumentation/internal/InstrumentationTestsActionFactory.kt
@@ -2,8 +2,8 @@ package com.avito.instrumentation.internal
 
 import com.avito.android.TestSuiteLoaderImpl
 import com.avito.android.runner.devices.DevicesProviderFactory
-import com.avito.android.runner.report.Report
 import com.avito.android.runner.report.ReportFactory
+import com.avito.android.runner.report.ReportFactoryImpl
 import com.avito.android.stats.StatsDSender
 import com.avito.http.HttpClientProvider
 import com.avito.instrumentation.internal.executing.TestExecutorFactory
@@ -29,10 +29,6 @@ internal interface InstrumentationTestsActionFactory {
 
         private val gson: Gson = Companion.gson
 
-        private val reportFactory: ReportFactory = params.reportFactory
-
-        private val report: Report = reportFactory.createReport()
-
         private val schedulerFactory: TestsSchedulerFactory
 
         private val finalizerFactory: FinalizerFactory
@@ -51,6 +47,17 @@ internal interface InstrumentationTestsActionFactory {
                 loggerFactory = params.loggerFactory
             )
 
+            val reportFactory: ReportFactory = ReportFactoryImpl(
+                timeProvider = timeProvider,
+                useInMemoryReport = params.useInMemoryReport,
+                buildId = params.buildId,
+                loggerFactory = params.loggerFactory,
+                httpClientProvider = httpClientProvider,
+                reportViewerConfig = params.reportViewerConfig
+            )
+
+            val report = reportFactory.createReport()
+
             this.schedulerFactory = TestsSchedulerFactory.Impl(
                 params = params,
                 report = report,
@@ -59,7 +66,8 @@ internal interface InstrumentationTestsActionFactory {
                 testExecutorFactory = TestExecutorFactory.Implementation(),
                 testSuiteLoader = TestSuiteLoaderImpl(),
                 timeProvider = timeProvider,
-                httpClientProvider = httpClientProvider
+                httpClientProvider = httpClientProvider,
+                reportFactory = reportFactory
             )
 
             this.finalizerFactory = FinalizerFactory.Impl(
