+++ subprojects/gradle/instrumentation-tests/src/main/kotlin/com/avito/instrumentation/internal/finalizer/AvitoReportViewerFinishAction.kt
@@ -1,19 +1,18 @@
 package com.avito.instrumentation.internal.finalizer
 
-import com.avito.android.runner.report.AvitoReport
+import com.avito.android.runner.report.LegacyReport
 import com.avito.instrumentation.internal.TestRunResult
 import com.avito.instrumentation.internal.finalizer.InstrumentationTestActionFinalizer.FinalizeAction
-import com.avito.instrumentation.internal.report.HasNotReportedTestsDeterminer
 
 internal class AvitoReportViewerFinishAction(
-    private val avitoReport: AvitoReport
+    private val legacyReport: LegacyReport
 ) : FinalizeAction {
 
     override fun action(testRunResult: TestRunResult) {
         if (testRunResult.notReported is HasNotReportedTestsDeterminer.Result.HasNotReportedTests) {
             val lostTests = testRunResult.notReported.lostTests
-            avitoReport.sendLostTests(lostTests)
+            legacyReport.sendLostTests(lostTests)
         }
-        avitoReport.finish()
+        legacyReport.finish()
     }
 }
