+++ subprojects/gradle/instrumentation-tests/src/test/kotlin/com/avito/instrumentation/report/HasNotReportedTestsDeterminerTest.kt
@@ -1,6 +1,7 @@
 package com.avito.instrumentation.report
 
-import com.avito.instrumentation.internal.report.HasNotReportedTestsDeterminer
+import com.avito.instrumentation.internal.finalizer.HasNotReportedTestsDeterminer
+import com.avito.instrumentation.internal.finalizer.LegacyNotReportedTestsDeterminer
 import com.avito.report.model.SimpleRunTest
 import com.avito.report.model.TestStaticDataPackage
 import com.avito.report.model.createStubInstance
@@ -12,7 +13,7 @@ internal class HasNotReportedTestsDeterminerTest {
 
     @Test
     fun `determine - AllTestsReported - when all tests found in report`() {
-        val result = HasNotReportedTestsDeterminer.Impl()
+        val result = LegacyNotReportedTestsDeterminer()
             .determine(
                 runResult = listOf(
                     SimpleRunTest.createStubInstance(name = "com.Test.test1", deviceName = "api22"),
@@ -31,7 +32,7 @@ internal class HasNotReportedTestsDeterminerTest {
 
     @Test
     fun `determine - ThereWereMissedTests - when not all tests found in report`() {
-        val result = HasNotReportedTestsDeterminer.Impl()
+        val result = LegacyNotReportedTestsDeterminer()
             .determine(
                 runResult = listOf(
                     SimpleRunTest.createStubInstance(name = "com.Test.test1", deviceName = "api22"),
@@ -45,20 +46,5 @@ internal class HasNotReportedTestsDeterminerTest {
             )
 
         assertThat(result).isInstanceOf<HasNotReportedTestsDeterminer.Result.HasNotReportedTests>()
-    }
-
-    @Test
-    fun `determine - FailedToGetMissedTests - when failed to get tests from report`() {
-        val result = HasNotReportedTestsDeterminer.Impl()
-            .determine(
-                runResult = emptyList(),
-                allTests = listOf(
-                    TestStaticDataPackage.createStubInstance(name = "com.Test.test1", deviceName = "api22"),
-                    TestStaticDataPackage.createStubInstance(name = "com.Test.test2", deviceName = "api22"),
-                    TestStaticDataPackage.createStubInstance(name = "com.Test.test3", deviceName = "api22")
-                )
-            )
-
-        assertThat(result).isInstanceOf<HasNotReportedTestsDeterminer.Result.DetermineError>()
     }
 }
