+++ subprojects/gradle/instrumentation-tests/src/test/kotlin/com/avito/instrumentation/suite/filter/StubFilterFactory.kt
@@ -1,5 +1,6 @@
 package com.avito.instrumentation.suite.filter
 
+import com.avito.android.runner.report.ReportFactory
 import com.avito.android.runner.report.StubReportFactory
 import com.avito.instrumentation.configuration.InstrumentationFilter
 import com.avito.instrumentation.createStub
@@ -9,17 +10,18 @@ import com.avito.instrumentation.internal.suite.filter.ImpactAnalysisResult
 import com.avito.logger.LoggerFactory
 import com.avito.logger.StubLoggerFactory
 
-internal object FilterFactoryFactory {
+internal object StubFilterFactory {
 
     fun create(
         filter: InstrumentationFilter.Data = InstrumentationFilter.Data.createStub(),
         impactAnalysisResult: ImpactAnalysisResult = ImpactAnalysisResult.createStubInstance(),
-        loggerFactory: LoggerFactory = StubLoggerFactory
+        loggerFactory: LoggerFactory = StubLoggerFactory,
+        reportFactory: ReportFactory = StubReportFactory()
     ): FilterFactory {
         return FilterFactory.create(
             filterData = filter,
             impactAnalysisResult = impactAnalysisResult,
-            reportFactory = StubReportFactory,
+            reportFactory = reportFactory,
             loggerFactory = loggerFactory
         )
     }
