+++ subprojects/gradle/runner/report/src/main/kotlin/com/avito/android/runner/report/ReportFactory.kt
@@ -2,14 +2,13 @@ package com.avito.android.runner.report
 
 import com.avito.report.ReportLinkGenerator
 import com.avito.report.TestSuiteNameProvider
+import java.io.Serializable
 
-public interface ReportFactory {
+public interface ReportFactory : Serializable {
 
     public fun createReport(): Report
 
-    public fun createReadReport(): ReadReport
-
-    public fun createAvitoReport(): AvitoReport
+    public fun createAvitoReport(): LegacyReport
 
     public fun createReportLinkGenerator(): ReportLinkGenerator
