+++ subprojects/gradle/runner/report/src/test/kotlin/com/avito/android/runner/report/TestReporterTest.kt
@@ -16,7 +16,7 @@ internal class TestReporterTest {
     @Test
     fun `every lost tests reported by separate request when batch size is 1`() {
         val reportsApi = StubReportsApi()
-        val reporter = AvitoReport.createStubInstance(
+        val reporter = LegacyReport.createStubInstance(
             reportsApi = reportsApi,
             batchSize = 1,
             buildId = buildId
